pyproject.toml
[project]
name = "agenticlangraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.3.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.4",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.4.8",
    "langsmith>=0.4.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]
